Fix old data structures

+++ MPImage/UIImage+MPAdditions.h
@@ -17,7 +17,7 @@
 - (UIImage *)imageByRemovingColorsWithMinColor:(uint)minColor maxColor:(uint)maxColor;
 - (UIImage *)imageByReplacingColor:(uint)color withColor:(uint)newColor;
 - (UIImage *)imageByReplacingColorsWithMinColor:(uint)minColor maxColor:(uint)maxColor withColor:(uint)newColor;
-- (UIImage *)imageByReplacingColorsWithMinColor:(uint)minColor maxColor:(uint)maxColor withColor:(uint)newColor andAlpha:(float)alpha;
+- (UIImage *)imageByReplacingColorsWithMinColor:(uint)minColor maxColor:(uint)maxColor withColor:(uint)newColor andAlpha:(CGFloat)alpha;
 
 - (UIImage *) flattenGrayscaleImage;
 @end
